requirements.txt
# Presentation Generator 依赖列表

# ========================================
# 核心依赖
# ========================================
google-genai>=1.0.0                   # Gemini API (新版)
zhipuai>=2.0.0                        # GLM API
openai>=1.0.0                         # OpenRouter API (第三备选)
python-dotenv>=1.0.0                  # 环境变量管理
Pillow>=10.0.0                        # 图像处理
requests>=2.31.0                      # HTTP 请求

# ========================================
# 可选依赖 (视频功能)
# ========================================
# ffmpeg-python>=0.2.0                 # FFmpeg Python 绑定 (可选)
